Move MirrorAction to mirror_service.py
