+++ .mypy.ini
@@ -87,6 +87,7 @@ modules =
     azul.plugins.metadata.anvil.service.filter,
     azul.plugins.metadata.hca.service.filter,
     azul.indexer.transform,
+    azul.indexer.mirror_service,
 
 packages =
     azul.openapi
