+++ src/azul/indexer/mirror_service.py
@@ -1,0 +1,15 @@
+from enum import (
+    auto,
+)
+
+from azul.queues import (
+    Action,
+)
+
+
+class MirrorAction(Action):
+    mirror_source = auto()
+    mirror_partition = auto()
+    mirror_file = auto()
+    mirror_part = auto()
+    finalize_file = auto()
